pyproject.toml
[project]
name = "yolo-trainer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "torchaudio>=2.2.0",
    "jupyterlab>=4.4.7",
    "ultralytics>=8.3.223",
    "patched-yolo-infer>=1.3.8",
    "tqdm>=4.67.1",
    "pillow>=11.3.0",
    "imageio>=2.37.0",
    "imageio-ffmpeg>=0.6.0",
    "sahi>=0.11.36",
    "pycocotools>=2.0.10",
    "pyyaml>=6.0.2",
]

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu124" },
]
torchvision = [
    { index = "pytorch-cu124" },
]
torchaudio = [
    { index = "pytorch-cu124" },
]
